New Items Added
-> Pause menu
-> New sound buttons on Pause Menu

+++ Assets/_Scripts/Google Play Gaming Services 1/VGPGSManager.cs
@@ -9,7 +9,7 @@ namespace Runtime
 {
     [Serializable]
     public class VPlayerData {
-        public long highestScore = 0;
+        public int highestScore = 0;
         public int level = 1;
         public int experience = 0;
         public bool[] achievements;
